Improve bullet lists.
Patch by Stuart Dickson

"
Here's the latest patch for the bullets functionality.

This includes the testdoc subfolder and there are two documents for testing the nested docs, one in LO and the other in Calligra.

The LO authored doc runs fine in Chrome, but on Firefox, the continued numbering has issues. It resets with each list, I believe the id-based selector may not be getting applied correclty, I'll add some additional decoration to the bullet block to confirm whether this is the case.

The Calligra document does not render correctly, which I was lookoing at before hitting build environment issues, which have taken my time the last couple of days (although look to be at an end now).

The things I wasn't too sure/happy with were:

    Having to copy code from Style2CSS into OdfCanvas (would prefer to have exposed the function and reused.
    I needed to get hold of the document in order to read the already added styles - and used window.document there (this works with jslint, but I'm not convinced it is the right approach)
"