Clarified error message when unavailable module is loaded
Previously, attempting to load a module with no manifest.js defined
for the directory the script is executed in would result in node throwing
an "Error: ENOENT, no such file or directory".

With these changes, node will report "Cannot load manifest for …" instead,
allowing it to then fail with the specific module being searched for. This makes
hunting down ordering problems in the compilation a little easier.