Modified check right confirmation to match check left
This will change the scan-right behaviour for valid cursor positions to
allow a cursor to be placed after trailing whitespace in a grouping
element provided there are more following characters in the paragraph.

According to positions in README_cursorposition, it is valid to place a
cursor after a whitespace element if there is another character in the
paragraph/heading element to the right. In the case of scanning to the
right for another valid character, TextPositionFilter.checkLeftRight
was only looking for siblings, and was not properly taking into account
grouping elements.