Don't reprocess each new space element during space expansion
getElementsByTagName* returns a HTML collection which is a live view
of the DOM tree (see MDN). As such, every new space created by the
expandSpaceElement sub-function adds an extra entry into this list,
and will be reprocessed (including clearing & replacing text node children).

Only the original space elements need to be processed, so the HTMLCollection
has been replaced with a fixed array instead.