Formatting.getAppliedStyles copes more gracefully with unknown style elements
The general use-case of getAppliedStyles will cope ok with missing entries
from the style sets. It makes dependent functions a little incorrect
if an unsupported style element is used, but as these callers are usually
checking for the presence of a particular format, this should rarely
lead to incorrect behaviour.