Always wrap styled text within a stylable container (text:span)
This ensures if the text node is within a non-stylable
element, the applicator first wraps the text in a container that can be
styled. This change will have no impact if the text is already in a
text:span as it usually is. This also ensures text properly styles
text:a content.

Note, the styling behaviour within text:a differs from LO significantly.
LibreOffice actually splits the bar text:a into multiple elements.
e.g.,:
<text:a>http://</text:a><text:span><text:a>webodf.org</text:a></text:span>

This patch will perform this styling as:
<text:a>http://<text:span>webodf.org</text:span></text:a>