Resynchronize the cursor position on a right-click selection (OSX only issue)
Right-clicking on Safari or Chrome on OSX causes the context menu to
appear, but the cursor is not properly updated to reflect the selection.
This means that the user could right-click & paste (for example), and
the text would appear where the cursor last was, rather than where
the right-click occurred. This problem does not occur on Windows.

This patch still is unable to expand to word-selection properly in the
browsers that try and do this as the current word expansion behaviour is
still broken in some circumstances.

TODO:
-Recreate browser default behaviours on right click (e.g., expand to word boundary on OSX)