Properly update odf cursor when mouse released ouside canvas
click and drag over a scetion of text to select it, ensuring to release the mouse-click outside of the editor. the section of text highlights but cursor remains where it was previously.
this is because the canvas element does not receive the mouse up event when releasing the mouse-click outside the canvas

changes include:
* listen to the window object rather than canvas element
* if anchor node or foucs node is outside the canvas, replace it with the closest node within the document
* ensure selectWord/selectParagraph is not called when user clicks outside of canvas