Only update the user's selection if the mouse down occurs within the canvas
This fixes a regression introduced in f7e9c89. Binding to the global
window's mouse up is necessary to allow the user selection to update.
The change introduced an undesirable side-effect though that caused clicks
entirely outside the odf canvas to reset the user selection to the
first or last position in the document.