Replace click handler with equivalent event.detail property
As per http://www.w3.org/TR/DOM-Level-3-Events/#event-type-mouseup,
event.detail records the number of clicks in the same way ClickHandler
was attempting to.

This has the happy side-effect of fixing an issue where double or triple
clicking on the editor's toolbar would result in the selection expanding
to word or paragraph boundaries.