Clarified SelectionMover.countStepsToPosition definition
countStepsToPosition now returns the difference in document positions
between the cursor and supplied point. This is slightly different to the
previous behaviour that was returning the number of valid positions
between the two points. This impacts behaviour when counting backwards
to an unworkable position by increasing the returned step count by one.