Allow double-click drag selection changes
Previously, the selection was being reset back to the word-boundary
of the word that contained the cursor at the time of the double-click.

Now, the browser selection is always updated, with additional logic to
expand the selection to the nearest boundary based on the double or
triple-click status of the event.