Add functions isAnnotation and isAnnotationWrapper to OdfUtils.
Add a function isInlineRoot to OdfUtils, which
indicates whether an element constitutes a 'root'
as respected by the RootFilter. In gui-less mode,
<office:annotation> is such a root, and in gui
mode, where we wrap annotations, we should also
treat the div with class "annotationWrapper" as
such a root.

Inline Roots are to be treated just like character
elements or frames with 'as-char', therefore add
a check for those to isCharacterElement as well,
alongside the one for isCharacterFrame.