Use operation groups to improve undo behaviour
Support has been added for operation groups. Operation groups (i.e.,
a group of ops queued together) are never split across multiple undo
states.

When groups are used, a new operation will be considered continuous
with the current undo state if it is continuous with the most recent
operation of the same type in the most recent group in the existing
undo state.