Fix toolbar positioning on body scroll.
When, after clicking in the text, the iOS keyboard
pops up, the entire editor body is scrolled upwards
because the window is shrunk, even if there was
no way to scroll it before.
This means that the toolbar slides upwards and out
of the visible part of the viewport.

This is unacceptable, therefore this commit introduces
a fix:
CSS-translate the toolbar when detecting a scroll. focusout,
or touchmove event.