Cope with absolute file destinations on Windows
It is assumed that absolute file sources should be copied to the root of
the target directory.

On unix-based systems, an absolute path begins with a '/', which is
interpreted as 'start at root' and does the right thing. On Windows,
these paths contain ':' (e.g., C:/Source) which is an invalid path.