Add StepCache support for multiple bookmarks with the same step number
It is valid for multiple bookmarks to share the same step number if
the user of the steps cache does not have a perfectly granular position
filter.

For example, paragraphs within a tracked changes block are not valid
steps according to the TextPositionFilter. This means all these
paragraphs stack up on the same step.