Add workarounds for qtruntimetest's version of WebKit
Content needs to be on a new line if it contains slashes due to a bug in
older versions of webkit. E.g., the one used in the qt runtime tests.
See https://bugs.webkit.org/show_bug.cgi?id=35010

In addition, asking for the value of a content block that contains a
counter returns a strange counter string such as "counter(-879878e564)".
I can't find a bug report for this, but it only occurs inside the qt
runtime tests.