Ensure merged paragraph style names are well defined
Though getAttributeNS is now standardised as of HTML4 to always return
null, according to MDN documentation browsers can return a variety of
values for a non-present attribute (generally null or ""). Update the
logic in TextController to always use "" for the merged style name as
the inline comment suggests is the expected behaviour.