Remove unnecessary (and impossible to use) constraint argument
The constraint function only allows a choice between two adjacent steps
if the supplied position is not exactly on a step. With things such as
inline roots and other complex features of the addressing system, this
is extremely difficult to use correctly.

As it is currently unused, the best thing to prevent future problems it
to terminate it now!